+++ internal/models/models.go
@@ -12,17 +12,17 @@ type (
 	}
 
 	StorageURL struct {
-		UserID      uint64 `json:"userID"`
+		UserID      uint32 `json:"userID"`
 		ShortURL    string `json:"short_url"`
 		OriginalURL string `json:"original_url"`
 	}
 
 	Storage interface {
-		Save(ctx context.Context, url1, url2 string, userID uint64) error
+		Save(ctx context.Context, url1, url2 string, userID uint32) error
 		GetOriginURL(ctx context.Context, shortURL string) (string, error)
 		GetShortURL(ctx context.Context, longURL string) (string, error)
 		Ping(ctx context.Context) error
-		GetUrlsUser(ctx context.Context, userID uint64) ([]StorageURL, error)
+		GetUrlsUser(ctx context.Context, userID uint32) ([]StorageURL, error)
 	}
 
 	Batch struct {
