+++ internal/storages/storage/storageurls.go
@@ -32,7 +32,7 @@ func (s *Storage) GetShortenURL(ctx context.Context, inputURL string) (string, e
 	return s.storage.GetShortURL(ctx, inputURL)
 }
 
-func (s *Storage) SaveURL(ctx context.Context, url1, url2 string, userID uint64) error {
+func (s *Storage) SaveURL(ctx context.Context, url1, url2 string, userID uint32) error {
 	return s.storage.Save(ctx, url1, url2, userID)
 }
 
@@ -40,6 +40,6 @@ func (s *Storage) Ping(ctx context.Context) error {
 	return s.storage.Ping(ctx)
 }
 
-func (s *Storage) GetUrlsUser(ctx context.Context, userID uint64) ([]models.StorageURL, error) {
+func (s *Storage) GetUrlsUser(ctx context.Context, userID uint32) ([]models.StorageURL, error) {
 	return s.storage.GetUrlsUser(ctx, userID)
 }
